+++ app/src/main/AndroidManifest.xml
@@ -15,18 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.First_week_mission"
         tools:targetApi="31">
-        <activity
-            android:name=".ui.detail.PokemonDetailActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.MainActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
     </application>
 
 </manifest>
