+++ feature/main/src/main/res/drawable/favorite_24dp.xml
