+++ feature/main/src/main/res/drawable/favorite_border_24dp.xml
