+++ feature/main/src/main/res/layout/activity_pokemon_detail.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.detail.PokemonDetailActivity">
+    tools:context=".detail.PokemonDetailActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
