+++ feature/main/src/main/res/layout/item_pokemon_info.xml
@@ -7,7 +7,7 @@
     <TextView
         android:id="@+id/tv_content"
         android:textSize="14sp"
-        android:textColor="@color/black"
+        android:textColor="#000000"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:padding="16dp"/>
