+++ feature/main/src/main/res/menu/detail_menu.xml
